refactor: split app planner label helpers
